budgetportal/requirements.txt
arrow==0.15.5
backports.csv==1.0.7
backports.functools-lru-cache==1.6.1
blessed==1.17.2
boto3==1.12.21
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
ckanapi==4.3
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.20
django-admin-sortable==2.2.3
django-adminplus==0.5
django-allauth==0.41.0
django-autoslug==1.9.6
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-extensions==2.2.8
django-environ==0.4.5
django-filter==2.2.0
django-haystack==2.8.1
django-import-export==2.0.2
django-js-asset==1.2.2
django-markdownify==0.8.0
django-partial-index==0.6.0
django-picklefield==2.1.1
django-pipeline==1.7.0
django-q==1.1.0
django-recaptcha==2.0.6
djangorestframework==3.12.4
djangorestframework-csv==2.1.0
django-storages==1.9.1
docopt==0.6.2
drf-haystack==1.8.9
et-xmlfile==1.0.1
futures==3.1.1
gevent==1.4.0
gunicorn==20.0.4
httplib2==0.19.1
idna==2.8
jdcal==1.4.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pathlib==1.0.1
psycopg2==2.8.4
pyaml==19.12.0
pysolr==3.8.1
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.4.1
requests==2.22.0
requests-oauthlib==1.3.0
sentry-sdk==0.14.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
tablib==1.1.0
Unidecode==1.1.1
urllib3==1.25.8
wagtail==2.7.4
wcwidth==0.1.8
Werkzeug==1.0.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
